backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
opencv-python-headless==4.11.0.86
numpy==2.2.3
Pillow==11.1.0
tflite-runtime==2.14.0; platform_machine != "x86_64"
zeroconf==0.146.1
